src/test/java/com/jparams/object/builder/provider/FloatProviderTest.java
package com.jparams.object.builder.provider;

import com.jparams.object.builder.type.Type;

import org.junit.Before;
import org.junit.Test;

import static org.assertj.core.api.Assertions.assertThat;

public class FloatProviderTest
{
    private Provider subject;

    @Before
    public void setUp()
    {
        subject = new FloatProvider();
    }

    @Test
    public void testSupports()
    {
        assertThat(subject.supports(Type.forClass(Float.class))).isTrue();
        assertThat(subject.supports(Type.forClass(float.class))).isTrue();
        assertThat(subject.supports(Type.forClass(String.class))).isFalse();
    }

    @Test
    public void testProvide()
    {
        assertThat(subject.provide(null)).isInstanceOf(Float.class);
    }
}